Moving calendar on the new timer module
