+++ api/kogito-api/src/main/java/org/kie/kogito/rules/RuleUnitInstance.java
@@ -18,7 +18,7 @@ package org.kie.kogito.rules;
 import java.util.List;
 import java.util.Map;
 
-import org.kie.api.time.SessionClock;
+import org.kie.kogito.timer.SessionClock;
 
 public interface RuleUnitInstance<T extends RuleUnitData> {
