+++ api/kogito-timer/src/main/java/org/kie/kogito/timer/SessionClock.java
@@ -14,13 +14,22 @@
  *  limitations under the License.
  */
 
-package org.kie.api.time;
+package org.kie.kogito.timer;
 
+/**
+ * A clock interface that all engine clocks must implement
+ */
+public interface SessionClock {
 
+    /**
+     * Returns the current time. There is no semantics attached
+     * to the long return value, so it will depend on the actual
+     * implementation. For instance, for a real time clock it may be
+     * milliseconds.
+     *
+     * @return The current time. The unit of the time, depends on
+     * the actual clock implementation.
+     */
+    long getCurrentTime();
 
-public interface Calendars {
-    Calendar get(String identifier);
-
-    void set(String identifier,
-             Calendar value);
 }
