+++ drools/drools-compiler/src/test/java/org/drools/compiler/integrationtests/FirstOrderLogicTest.java
@@ -52,8 +52,8 @@ import org.kie.api.runtime.KieSession;
 import org.kie.api.runtime.KieSessionConfiguration;
 import org.kie.api.runtime.conf.ClockTypeOption;
 import org.kie.api.runtime.rule.FactHandle;
-import org.kie.api.time.SessionClock;
-import org.kie.api.time.SessionPseudoClock;
+import org.kie.kogito.timer.SessionClock;
+import org.kie.kogito.timer.SessionPseudoClock;
 import org.kie.internal.builder.KnowledgeBuilder;
 import org.kie.internal.builder.KnowledgeBuilderFactory;
 import org.kie.internal.io.ResourceFactory;
