+++ drools/drools-compiler/src/test/java/org/drools/compiler/integrationtests/WindowTest.java
@@ -36,7 +36,7 @@ import org.kie.api.runtime.KieSession;
 import org.kie.api.runtime.KieSessionConfiguration;
 import org.kie.api.runtime.conf.ClockTypeOption;
 import org.kie.api.runtime.rule.EntryPoint;
-import org.kie.api.time.SessionPseudoClock;
+import org.kie.kogito.timer.SessionPseudoClock;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
