+++ drools/drools-compiler/src/test/java/org/drools/compiler/integrationtests/marshalling/MarshallingTest.java
@@ -96,8 +96,8 @@ import org.kie.api.runtime.conf.ClockTypeOption;
 import org.kie.api.runtime.conf.TimedRuleExecutionOption;
 import org.kie.api.runtime.conf.TimerJobFactoryOption;
 import org.kie.api.runtime.rule.EntryPoint;
-import org.kie.api.time.SessionClock;
-import org.kie.api.time.SessionPseudoClock;
+import org.kie.kogito.timer.SessionClock;
+import org.kie.kogito.timer.SessionPseudoClock;
 import org.kie.internal.builder.KnowledgeBuilderConfiguration;
 import org.kie.internal.builder.KnowledgeBuilderFactory;
 import org.kie.internal.marshalling.MarshallerFactory;
