+++ drools/drools-core/src/main/java/org/drools/core/command/GetSessionClockCommand.java
@@ -19,7 +19,7 @@ package org.drools.core.command;
 import org.kie.api.command.ExecutableCommand;
 import org.kie.api.runtime.Context;
 import org.kie.api.runtime.KieSession;
-import org.kie.api.time.SessionClock;
+import org.kie.kogito.timer.SessionClock;
 import org.kie.internal.command.RegistryContext;
 
 public class GetSessionClockCommand
