+++ drools/drools-core/src/main/java/org/drools/core/command/runtime/GetCalendarsCommand.java
@@ -17,7 +17,7 @@
 package org.drools.core.command.runtime;
 
 import org.kie.api.command.ExecutableCommand;
-import org.kie.api.time.Calendars;
+import org.kie.kogito.timer.Calendars;
 import org.kie.api.runtime.Context;
 import org.kie.api.runtime.KieSession;
 import org.kie.internal.command.RegistryContext;
