+++ drools/drools-core/src/main/java/org/drools/core/command/runtime/GetSessionTimeCommand.java
@@ -26,7 +26,7 @@ import org.drools.core.runtime.impl.ExecutionResultImpl;
 import org.kie.api.command.ExecutableCommand;
 import org.kie.api.runtime.Context;
 import org.kie.api.runtime.KieSession;
-import org.kie.api.time.SessionClock;
+import org.kie.kogito.timer.SessionClock;
 import org.kie.internal.command.RegistryContext;
 
 @XmlRootElement(name="get-session-time")
