+++ drools/drools-core/src/main/java/org/drools/core/phreak/PhreakTimerNode.java
@@ -56,7 +56,7 @@ import org.drools.core.time.impl.Timer;
 import org.drools.core.util.LinkedList;
 import org.drools.core.util.index.TupleList;
 import org.kie.api.definition.rule.Rule;
-import org.kie.api.time.Calendars;
+import org.kie.kogito.timer.Calendars;
 import org.kie.api.runtime.conf.TimedRuleExecutionFilter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
