+++ drools/drools-core/src/main/java/org/drools/core/time/SessionPseudoClock.java
@@ -16,5 +16,5 @@
 
 package org.drools.core.time;
 
-public interface SessionPseudoClock extends org.kie.api.time.SessionPseudoClock {
+public interface SessionPseudoClock extends org.kie.kogito.timer.SessionPseudoClock {
 }
