+++ drools/drools-core/src/main/java/org/drools/core/time/impl/ExpressionIntervalTimer.java
@@ -34,7 +34,7 @@ import org.drools.core.spi.Tuple;
 import org.kie.kogito.timer.Trigger;
 import org.kie.kogito.timer.impl.DefaultJobHandle;
 import org.kie.kogito.timer.impl.IntervalTrigger;
-import org.kie.api.time.Calendars;
+import org.kie.kogito.timer.Calendars;
 
 public class ExpressionIntervalTimer  extends BaseTimer
     implements
