+++ drools/drools-core/src/main/java/org/drools/core/time/impl/Timer.java
@@ -25,7 +25,7 @@ import org.drools.core.spi.RuleComponent;
 import org.drools.core.spi.Tuple;
 import org.kie.kogito.timer.Trigger;
 import org.kie.kogito.timer.impl.DefaultJobHandle;
-import org.kie.api.time.Calendars;
+import org.kie.kogito.timer.Calendars;
 
 public interface Timer extends Serializable, RuleComponent, RuleConditionElement {
